global-style.css, App.css and index.css moved to style folder.
